Update volunteerForm.html
Ran through validator. Needed to do something weird for the select for the security question. it was an error that said: "The first child option element of a select element with a required attribute, and without a multiple attribute, and without a size attribute whose value is greater than 1, must have either an empty value attribute, or must have no text content. Consider either adding a placeholder option label, or adding a size attribute with a value equal to the number of option elements." So I found a solution.